Avoid prematurely logging user in (before token has been verified).
This makes sure that no other plugin is able to authenticate the
user by consuming the credentials we receive - we verify those ourselves
and only want the user to be logged in *after* we checked the token.

Before this change the following happened:
- The GoogleAuthenticatorPlugin returns `None` after verifying
  credentials.
- Therefore PluggableAuthService._extractUserIds() tries the next
  IAuthenticatorPlugin in the list, until it finds one that is able
  to successfully authenticate the credentials.
- Because whichever plugin can authenticate those credentials then
  returns a not-None value upon success, the user is actually
  logged in, before we got a chance to verify the token.
- Because of that, the user's auth session was destroyed in the
  token validation view by removing the __ac cookie.

This commit changes this behavior so the user never actually gets
logged in before we verified the token. We do this by consuming the
credentials that get passed into our authenticator plugin, so that
no later plugins can use them to authenticate the user prematurely.

Because that means the first attempt at authentication will fail
from PAS' point of view, there will be a "Login failed" status
message produced which we will need to remove in the token validation
view.

Destroying the user's auth session by deleting the __ac cookie is not
necessary any more either.